+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Switch to jakarta annotations (jakarta.annotation instead javax.annotation)
+
+NOTE: guice still use javax.inject
+
 ### 1.4.0 (2021-09-06)
 * Drop java 1.6 and 1.7 support
 * Update to guice 5.0.1